pyproject.toml
[project]
name = "researcher-zero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.6",
    "networkx>=3.4.2",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "tavily>=1.1.0",
    "tavily-python>=0.7.21",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["core"]
